+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+**Entity Action Types**
+- Added `direction` and `space` fields to `apugli:raycast` and `apugli:explosion_raycast`.
+
+**Entity Condition Types**
+- Added `direction` and `space` fields to `apugli:raycast`.
+
 **Optimisations**
 - Rewrote `apugli:cached_block_in_radius` from the ground up for optimisation reasons.